feat: cypress login e2e test (#449)
* feat: cypress login e2e test

* docs: readme pwa 설명 추가

* chore: Alarm interval 상수선언

* refactor: / 로 접근을 로그인을 취급

* chore: 기존 login 페이지 삭제.

* feat: add login command to Cypress
