+++ .github/workflows/cypress.yml
@@ -14,11 +14,11 @@ jobs:
       - name: Install dependencies
         run: yarn
 
-      # - name: Generate environment json file
-      #   run: |
-      #     echo "$CYPRESS_ENV_JSON" >> cypress.env.json
-      #   env:
-      #     CYPRESS_ENV_JSON: ${{ secrets.CYPRESS_ENV_JSON }}
+      - name: Generate environment json file
+        run: |
+          echo "$CYPRESS_ENV_JSON" >> cypress.env.json
+        env:
+          CYPRESS_ENV_JSON: ${{ secrets.CYPRESS_ENV_JSON }}
 
       - name: E2E test - chrome
         uses: cypress-io/github-action@v4
