+++ cypress.config.ts
@@ -1,11 +1,14 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { defineConfig } from 'cypress';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 export default defineConfig({
   projectId: 'q4deia',
   e2e: {
     setupNodeEvents(on, config) {
-      // implement node event listeners here
+      return require('./cypress/plugins/index.js')(on, config);
     },
     baseUrl: 'http://localhost:3000/',
     specPattern: 'cypress/e2e/**/*.{js,jsx,ts,tsx}',
