+++ cypress/e2e/inbox.spec.ts
@@ -1,0 +1,9 @@
+/// <reference types="cypress" />
+
+export {};
+
+describe('inbox', () => {
+  before(() => {
+    1;
+  });
+});
