--- cypress/e2e/spec.cy.ts
@@ -1,5 +1,0 @@
-describe('empty spec', () => {
-  it('passes', () => {
-    cy.visit('https://example.cypress.io');
-  });
-});
