+++ cypress/support/index.ts
@@ -8,16 +8,10 @@ declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace Cypress {
     interface Chainable {
-      /**
-       * Custom command to select DOM element by data-cy attribute.
-       * @example cy.dataCy('greeting')
-       */
       dataCy(value: string): Chainable<JQuery<HTMLElement>>;
+      login(): () => void;
     }
   }
 }
 
 // cypress/support/index.ts
-Cypress.Commands.add('dataCy', (value) => {
-  return cy.get(`[data-cy=${value}]`);
-});
