+++ src/shared/constants/delay.ts
@@ -1,2 +1,2 @@
 export const TRANSITION_DELAY = 300 as const;
-export const CHECKALARM_INTERVAL = 60000 as const;
+export const CHECK_ALARM_INTERVAL = 60000 as const;
